refactor: move OnDemandWorker to utils
