+++ packages/mask/src/plugins/Polls/messages.ts
@@ -1,6 +1,6 @@
 import { identifier } from './constants'
 import { createPluginMessage, createPluginRPC } from '@masknet/plugin-infra'
-import { OnDemandWorker } from '../../web-workers/OnDemandWorker'
+import { OnDemandWorker } from '../../../utils-pure'
 import { AsyncCall, _AsyncVersionOf } from 'async-call-rpc'
 import { WorkerChannel } from 'async-call-rpc/utils/web/worker'
