+++ packages/mask/utils-pure/index.ts
@@ -3,3 +3,4 @@ export * from './assertion'
 export * from './memoize'
 export * from './hmr'
 export * from './crypto'
+export * from './OnDemandWorker'
